% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.imo.R
\name{combine.imo}
\alias{combine.imo}
\title{combines estimated from different vertical splitted sub-samples using appropriate combination rules}
\usage{
combine.imo(imo.param.est, imo.param.cov)
}
\arguments{
\item{imo.param.est}{a matrix containing estimated parameters in each sub-sample as its rows}

\item{imo.param.cov}{a list contaning the covariance matrix estimated within each sub-sample.}
}
\value{
the combined estimate vector and covariance matrix.
}
\description{
combines estimated from different vertical splitted sub-samples using appropriate combination rules
}
\author{
Vahid Nassiri
}
