% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imo.subsamp.R
\name{imo.subsamp}
\alias{imo.subsamp}
\title{take random sub-samples from each cluster using random vertical splitting with acceleration}
\usage{
imo.subsamp(data, num.outp, sub.size)
}
\arguments{
\item{data}{the input dataset}

\item{num.outp}{number of sub-samples}

\item{sub.size}{size of each sub-sample}
}
\value{
the sub-sampled dataset
}
\description{
take random sub-samples from each cluster using random vertical splitting with acceleration
}
\author{
Vahid Nassiri
}
