% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.mixed.data.prep.R
\name{multi.mixed.data.prep}
\alias{multi.mixed.data.prep}
\title{prepares a multivariate dataset
in a format suitable for joint modelling approach.}
\usage{
multi.mixed.data.prep(data, num.var)
}
\arguments{
\item{data}{could be the outcome of gen.multi.lin.mixed
  it should have the responses as columns, followed by
repeated measurements id and cluster id.}

\item{num.var}{number of responses}
}
\value{
the prepared dataset
}
\description{
prepares a multivariate dataset
in a format suitable for joint modelling approach.
}
\author{
Vahid Nassiri
}
