% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.mixed.pairs.prep.R
\name{multi.mixed.pairs.prep}
\alias{multi.mixed.pairs.prep}
\title{takes all sub-samples based on pairwise approach, i.e., 
all possible pairs of outcomes. It can be used in any package 
to fit the models.}
\usage{
multi.mixed.pairs.prep(data, num.var)
}
\arguments{
\item{data}{could be the outcome of multi.mixed.data.prep or any dataset with the same format}

\item{num.var}{number of response variables}
}
\value{
the prepared dataset
}
\description{
takes all sub-samples based on pairwise approach, i.e., 
all possible pairs of outcomes. It can be used in any package 
to fit the models.
}
\author{
Vahid Nassiri
}
